requirements.txt
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
PyMatting==1.1.13
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
tifffile==2024.9.20
tqdm==4.67.0
urllib3==2.2.3
Werkzeug==3.1.3
gunicorn==23.0.0
onnxruntime==1.20.1